+++ app/controllers/projects_controller.rb
@@ -5,7 +5,7 @@ class ProjectsController < ApplicationController
 
   def index
     authorize Project
-    @pagy, @projects = pagy(policy_scope(Project.open))
+    @pagy, @projects = pagy(policy_scope(Project.includes(:task_list).open))
 
     EstablishmentsFetcher.new.call(@projects)
     IncomingTrustsFetcher.new.call(@projects)
@@ -13,7 +13,7 @@ class ProjectsController < ApplicationController
 
   def completed
     authorize Project
-    @pagy, @projects = pagy(policy_scope(Project.completed))
+    @pagy, @projects = pagy(policy_scope(Project.includes(:task_list).completed))
   end
 
   def show