+++ db/migrate/20230131113754_add_confirmed_conversion_date_to_projects.rb
@@ -1,0 +1,5 @@
+class AddConfirmedConversionDateToProjects < ActiveRecord::Migration[7.0]
+  def change
+    add_column :conversion_voluntary_task_lists, :stakeholder_kick_off_confirmed_conversion_date, :date
+  end
+end